requirements.txt
# Core runtime dependencies
duckdb>=0.10
requests>=2.31
pyyaml>=6.0

# Development dependencies
pytest>=8.0
flake8>=7.0
